\name{myAnimalINLAfixes-package}
\alias{myAnimalINLAfixes-package}
\alias{myAnimalINLAfixes}
\docType{package}
\title{
my AnimalINLA fixes
}
\description{
Some updates of INLA function names that have changed, and that were not transposed to the AnimalINLA code.
}
\details{
\tabular{ll}{
Package: \tab myAnimalINLAfixes\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-10-01\cr
License: \tab GPL3.0\cr
}
}
\author{Facundo Muñoz <facundo.munoz@orleans.inra.fr>}
\references{http://www.r-inla.org/related-projects/animalinla}